refactor(sync): extract http2 fallback transport into its own file
Move http2FallbackTransport, isHTTP2FramingError, and the transport
construction out of service.go into pkg/extensions/sync/http2_fallback.go.
service.go now calls a single newHTTP2FallbackTransport() helper.

Same anti-pattern rchincha flagged on #4063 ("too much going in
pkg/api/routes.go wrt storage etc. We should move all that code out to
pkg/api/blobstream.go"). Pure refactor — no behavior change.
